requirements.txt
pymodbus==3.6.4
psycopg2-binary==2.9.9
